Fix Header-Intro component collapsing.
Components after the Header-Intro component would be displayed just below the background gradient. This is caused by tables that hold the email content running over the background div.

I found a fix by applying the background div to the main container element that will render the entire email, and having the entire email break out of this div to display.