Fix for editable buttons within MailChimp.
MailChimp would add overriding styling to a tags within editable buttons. The workaround for this was to implement my own HTML partial for a button, which can be re-purposed as a button for other components.

The styling for this a tag within the custom button would be included in a media query of greater than a pixel, as a hack for the  Foundation for emails framework to include it at build time in the head of the email - ready for import to MailChimp.