Fix Gmail styling inline links.
Gmail will style anchor tags blue. The fix for this is to add a class to the parent object that the link could sit within, in order to be able to reference it, and add custom styling for the links within those classed parents.

The media query is a hack that will get Mailchimp to inline the styles when sending, as there may or may not be a link inserted by the user.